+++ packages/yak_runner/lib/src/custom/errors/bool_check_false.dart
@@ -1,2 +1,0 @@
-/// an error to be throw when a bool check return false
-class BoolCheckFalse extends Error {}