+++ packages/yak_runner/lib/src/custom/errors/cannot_cast_as.dart
@@ -1,2 +1,0 @@
-/// an error to be throw when `Foo as Bar` fails
-class CannotCast extends Error {}