+++ packages/yak_runner/lib/src/custom/runners/type_as_type.dart
@@ -1,0 +1,11 @@
+import '../../all.dart';
+
+/// a `YakRunnerArg` that returns a T if contained in a given iterable
+class TypeAsType<T, S> extends YakRunnerArg<T, S> {
+  ///
+  TypeAsType()
+      : super(
+          (obj) => obj as T,
+          errorsWhitelist: {TypeError().ignore},
+        );
+}